Submit for stash 07

+++ log.go
@@ -20,9 +20,13 @@ func DebuglnRType(i string, rd int, rs1 int, rs2 int) {
 func DebuglnIType(i string, rd int, rs1 int, imm uint64) {
 	Debugln(fmt.Sprintf("Instr: % 10s | rd: 0x%02x rs1: 0x%02x imm: 0x%04x", i, rd, rs1, imm))
 }
+
 func DebuglnSType() {}
+
 func DebuglnBType() {}
-func DebuglnUType(i string, rd int, imm uint32) {
+
+func DebuglnUType(i string, rd int, imm uint64) {
 	log.Printf("Instr: % 10s | rd: 0x%02x imm: 0x%04x", i, rd, imm)
 }
+
 func DebuglnJType() {}
